Added google maps link
There's now a link to the google maps directions for each restaurant. Leave the commented out code in, I know it's a graveyard but it's there in case I can figure out/we need mapView.

That being said, mapView is a pain in the ass and even if we did put it in the observations page it would take up WAY too much screen space, so far the link is the best way to go.